// frontend/src/api/generated/models/bOMPayload.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface BOMPayload {
  product_variant_id: string;
  version?: number;
  effective_date: string;
  is_active?: boolean;
}